Completely refactor database implementation
The main new feature is a new `Entry`-type API which exposes the
`rusqlite::Transaction` to the caller. The `DatabaseEntry` is attached to a row
in the `Records` table, which either exists (in which case the
corresponding `Row` contains the relevant `RowId`), or it does not
exist.

These invariants - either the row existing or not - are then upheld for
the remainder of the life of the transaction. This enables the caller to
execute more logic which is conditional on the database being in a
particular state, without having to rely on the `RecordDatabse`
implementing the relevant methods.

This also simplifies a lot of return types, since the initial errors
involving existence/non-existence of a given row can be processed
immediately and then do not need to be handled for the remainder of the
transaction.

Along the way, this commit also implements new `update` and `delete`
commands

It also updates some of the statement caching to cache fewer
statements and improves return types to communicate result state to
the user more precisely.